move gpu_impl of elementwise kernel
