--- paddle/pten/kernels/cpu/CMakeLists.txt
@@ -1,1 +1,0 @@
-cc_library(math_cpu SRCS math.cc DEPS dense_tensor kernel_context kernel_factory eigen_function blas pten_transpose_cpu cast_kernel)
